requirements.txt
Pillow
aiohttp
numpy
